AllVerbsMixin: call a method for all HTTP verbs
The dispatch method will extra the attribute specified by the
class attribute "all_handler", the default value of which is "all".
The dispatch method then calls this method, without checking what
the HTTP verb is. If "all_handler" is not specified (techincally
if it evaluates to False) then ImproperlyConfigured exception is
raised.

Tests are in tests/test_other_mixins.py::TestAllVerbsMixin.