Fixed the tests (that prevented detection of the check_membership).
test_with_group_list
Previously, the test modified a group_required property from a GroupRequiredView **instance**, but not from the **class**. When the client is launched to perform a request it instantiates a **new** instance, with the original class variable 'test_groups' and not the intended sequence.
Now the test modified the class variable, so the client can create an instance with the correct attribute. (value is restored before finishing the test).

test_with_unicode
Similar bug with the instance variable vs class variable, which went undetected because of another bug when modifying the group name. When calling user.groups.all()[0] an object is returned (and then modified), but since its reference is not saved, the changes are lost (rerunning user.groups.all()[0] returns the instance saved in the DB, and not the modified one). That bug prevented the user group from being modified, so the test deceptively passed.