call redirect_url on LOGIN_REDIRECT_URL
django has a standard where on LOGIN_REDIRECT_URL settings it's
permitted to use url names like 'namespace:urlname' without using
reverse_lazy on it (given it could be troublesome to import that
directly in settings), so calling redirect_url(LOGIN_REDIRECT_URL)
fixes this problem. This calling standard is also done in the
django.contrib.auth.login view, so we should follow that standard.